requirements.txt
Flask>=1.0.2
GitPython>=2.1.11
requests>=2.19.1
